cleaner metadata appending & dedicated termination system with sweep rounds
metadata was until now just added via singular .pushback()s, in two different places, very brittle to any changes to metadata order or new addition. now it has its own function that automates this away

Sweeping can now be performed in multiple rounds, and the root node (and only the root node) keeps track of the number of rounds and publishes eventually a new termination signal to all other ranks. This is especially important for the case that new ranks join after some rounds have already finished, as their own round-counter would now be out-of-sync. To send the termination signal quickly to the child nodes, it is injected via a rootTransform at the root node