add instanct propagation of sweep round info
moved all the root-node operations to the actual (now in-place) root transformation. this means in particular that a new sweep round is directly injected the moment we notice it, and is no longer delayed by one full sharing period as before. Mallob tells the solvers immediately this new sweep round, so that they can adjust their internal environment sizes